+++ R/metacor.R
@@ -872,6 +872,10 @@ metacor <- function(cor, n, studlab,
       res$I2.resid <- res$lower.I2.resid <- res$upper.I2.resid <- NA
     }
     else {
+      res$Q.w.random <- hcc$Q.resid
+      res$df.Q.w.random <- hcc$df.Q.resid
+      res$pval.Q.w.random <- hcc$pval.Q.resid
+      ##
       res$tau2.resid <- hcc$tau2.resid
       res$lower.tau2.resid <- hcc$lower.tau2.resid
       res$upper.tau2.resid <- hcc$upper.tau2.resid
@@ -882,9 +886,9 @@ metacor <- function(cor, n, studlab,
       res$sign.lower.tau.resid <- hcc$sign.lower.tau.resid
       res$sign.upper.tau.resid <- hcc$sign.upper.tau.resid
       ##
-      res$Q.w.random <- hcc$Q.resid
-      res$df.Q.w.random <- hcc$df.Q.resid
-      res$pval.Q.w.random <- hcc$pval.Q.resid
+      res$Q.resid <- hcc$Q.resid
+      res$df.Q.resid <- hcc$df.Q.resid
+      res$pval.Q.resid <- hcc$pval.Q.resid
       ##
       res$H.resid <- hcc$H.resid
       res$lower.H.resid <- hcc$lower.H.resid