+++ NEWS.md
@@ -6,6 +6,9 @@
   heterogeneity variance in subgroups resulted in an error
   (bug was introduced in **meta**, version 4.16-0)
 
+* For GLMMs, export Wald-type Q statistic for residual heterogeneity
+  instead of missing value
+
 
 ## meta, version 4.19-1 (2021-09-14)